clean up build and fresh_start scripts

+++ scripts/build.sh
@@ -1,15 +1,31 @@
 #!/bin/sh
 
 if [ ! -e rcloud.support/DESCRIPTION ]; then
+    if [ -n "$ROOT" ]; then
+	echo "NOTE: changing to '$ROOT' according to ROOT"
+	cd "$ROOT"
+    fi
+fi
+
+if [ ! -e rcloud.support/DESCRIPTION ]; then
     echo '' 2>&1
     echo ' ERROR: cannot find rcloud.support. Please make sure you are' 2>&1
-    echo '        runnign this script form the RCloud root directory!' 2>&1
+    echo '        running this script from the RCloud root directory!' 2>&1
     echo '' 2>&1
     exit 1
 fi
 
-# build JS and rcloud.support
-make -C htdocs/js && R CMD build rcloud.support && RCS_SILENCE_LOADCHECK=TRUE R CMD INSTALL rcloud.support_`sed -n 's/Version: *//p' rcloud.support/DESCRIPTION`.tar.gz
+# build JS (if available)
+for dir in htdocs/js  htdocs/lib; do
+    if [ -d $dir/node_modules ]; then
+	echo " - checking JS code in $dir"
+	make -C $dir || exit 1
+    else
+	echo "   no node.js modules, skipping"
+    fi
+done
+
+R CMD build rcloud.support && RCS_SILENCE_LOADCHECK=TRUE R CMD INSTALL rcloud.support_`sed -n 's/Version: *//p' rcloud.support/DESCRIPTION`.tar.gz || exit 1
 
 # build internal packages (not in git)
 if [ -e internal ]; then
@@ -18,11 +34,13 @@ if [ -e internal ]; then
     done
 fi
 
+if [ -e ".git" ]; then
 # update branch/revision info
-REV=`( git rev-list --abbrev-commit -n 1 HEAD )`
-BRANCH=`( git status | sed -n 's:.*On branch ::p' | sed 's:/:-:g' )`
-
-if [ -n "$REV" ]; then
-    echo "$BRANCH" > REVISION
-    echo "$REV" >> REVISION
+    REV=`( git rev-list --abbrev-commit -n 1 HEAD )`
+    BRANCH=`( git status | sed -n 's:.*On branch ::p' | sed 's:/:-:g' )`
+    
+    if [ -n "$REV" ]; then
+	echo "$BRANCH" > REVISION
+	echo "$REV" >> REVISION
+    fi
 fi
